sys-app/portage: A real solution for ${EPREFIX}/var/lib/portage/config locking
Hallelujah!  This solution to the ${EPREFIX}/var/lib/portage/config access
denied problem is about a bazillion times better.

o Adds six bytes to portage sourcecode
o Uses existing functionality
o Not a kludge

now... if only I was certain it didn't defeat the purpose of the lock :(

Anyhow the solution is, ask portage to use a separate lockfile rather than
locking the file itself, which circumvents whetever is the weird cygwin bug
causing os.rename to fail to unlink the locked file.