profile.bashrc: stricter handling of PATH
Includes code to move /bin and /usr/bin to the end of PATH if present.
To generalize this code we should at least predicate it on EPREFIX being set
although this could be said of may other bits of profile.bashrc.

Also fix misc. bugs and annoyances in profile.bashrc