sys-apps/portage: cygdll_protect & cygwin_lib-qa-fix
This includes two things, first a fix for a QA failure triggered
by the new gen_usr_ldscript support -- once again the .dll.a is
showing up correctly as a static library but it is once again
incorrect to handle it like the .a

It also includes very experimental support for a new feature
which I'm calling cygdll_protect.  The idea is to mirror the
config_protect code in portage and use it to prevent merging of
.dll files which are loaded by portage at merge-time.  Once fully
implemented, this should fix the most vexing problems triggered by
dll-base changes and fork().  The code is not working yet (at least,
I don't think so...) so I have walled it off behind a useflag for now.