portage_workdir_hacktool
after loosing a shitload of work, I decided to get a bit more serious about the
pile of horrible scripts I'd cobbled together for hacking on workdirs in
/var/tmp/portage.  Considering what a gross kludge this is, it's really pretty
effective, at least for hacking on /var/tmp/portage/sys-apps/portage-*/work

The basic idea is to have a lot of quick-and-dirty commandline tools to roll up your
working changes into a patch in ${FILESDIR}, to keep some kind of
ultra-primitive revision history, and to unravel some of the patches-of-patches
muddle that results from version controlling ${FILESDIR} instead of
generating your patches from a real VCS (which, to be clear, is a way better solution,
but in my defense, inertia and slow, unthinking accretion brought me to this, not
any sort of plan.)

Not sure I really expect anybody else to use this, but FFS, caveat emptor, I will
not be held responsible for the massive data loss you may very well suffer if you
do (there's a bunch of rm -rf in there; seriously, be careful and exercise
some circumspection if you want to try this out).  Nevertheless, here it is
in case somebody has need of such a monstrosity.

You can treat it as public domain or GPL or whatever makes you feel warm and fuzzy.

-gmt