sys-apps/portage: noop: activate try/finally clause
This paves the way for subsequent use of the portage.master_lock module in portage
The fact that this changes a lot of lines, compounded with the fact that this repository
stores my changes as a diff-of-diffs makes this change really hellish.  By isolating it,
hopefully I can minimize that.  Really, if I'm going to hack on portage so much, I should
seriously think about a different approach but I'm pretty lazy and my scripts make it
awfully easy for /me/ to see what's going on :)

Sorry about that.

Anyhow.  To be clear: this change is a noop.  It mostly is adding tabs to emerge_main inbetween an
enormous

	try:
		.
		. (re-indented portage code)
		.
	finally:
		pass

construct.