sys-apps/portage: cygdll_protect: implement heuristic protections
This is the 100 lines patch I originally wanted to write at the
outset of the cygdll_protect vision quest.  Uses cygwin 'pldd' tool
to track down loaded executables for all running instances of portage
(using the handy master-lock database to find them).  Each executable
(under EPREFIX) is added to the _cygdllprotect_cruft_accumulator, which
is, in turn, fed into the master cygdll protections list.

A few odds and ends remain (it misses the excutable named on pldd's first
line, which is a bug, and -- maybe, I need to think about it -- there is
some missing code to update the protections under certain circumstances,
such as during package merges), but with this commit, the cygdll_protect
code should start serving its purpose and preventing portage crashes under
most previously intractable circumstances.

That being stated, obviously it's hot off the presses as of this commit, and
largely untested.