sys-apps/portage: cygwin-dosym-exe-hack
Implement an extension to the standard dosym so that naievely saying
dosym foo /bar will create a /bar.exe -> foo.exe symlink when, in fact,
/foo.exe is a cygwin executable and we can detect that (a) it's safe
to do so, and (b) cygwin is automagically hardlinking foo to foo.exe...
actually if there were a real hard-link there I guess the same thing
would happen, but I doubt that's a problem.

One issue that could thwart our heuristic is something like:

dosym quux /foo/bar/baz
dosym shazam /foo/bar/quux

where shazam is an executable.  The quux->shazam link would be created
as quux.exe to shazam.exe; however, the baz->quux link would be created
without the .exe since, as of when we call dosym, it doesn't meet the
criteria for automatic conversion.  To fix this we would have to keep a
dictionary of symlinks somewhere, or agressively modify recursive links
in response to a single dosym; that could get pretty silly, considering
that without going into the Windows filesystem its (almost) impossible to
tell whether the real filesystem object is baz or baz.exe.

Consider, what if baz and quux are installed by different packages!?  Then
if we want to magically do the "right" thing (a) what that is isn't entirely
clear and (b) it could get pretty complicated if we are trying to be
anal about it.

Where this distinction starts to become important is when we start testing
"does portage contain abc"?  At the moment, this will return false for
abc.exe; something probably needs to be done about that eventually since
for all intents and purposes portage installed abc and abc.exe as hard-links.

Anyhow, baby steps :)

Another question we could legitimately ask is: is portage better off without
this feature or with it?  I think the answer is probably that it's better off
with it but we'll see.