app-arch/bzip: reorganize
In addition to a version bump, this version brings the delta vs.
portage down to just a handful of lines, especially in -r3.

Additionally, this version does a lot less kludging around and
acts more like its supposed to.  That being stated it still kind
of feels like a mess.  But a much smaller mess.