sys-devel/gettext: deep prefixification
Sadly, it seems this will be the first of several patches like this.
Although cygdll-protect is working, a new layer of serious problems
has now been uncovered: when (non-prefixed) /bin/{ba,}sh is invoked
as a subshell of ${EPREFIX}/bin/bash, cygwin is spazzing out.

We need to supress this as much as possible.  Of course we can do
whatever on a per-package basis, but right now the most insidious
culprits are in ${EPREFIX}/usr/share/*/*.m4.  This is because,
a lot of the time, whetever measures we may take against /bin/sh,
autotools brings it right back from the dead.

So, we need to identify core packages and start thoroughly
delousing them.  This will be slow and painful but the end
result will hopefully justify the effort and spare us from
inelegant hacks or "rm /bin/bash".

I'm also hoping that these patches will start allowing us to pass
more tests -- I'm finding too many test failures are caused by
rebasing errors which creates a difficult circumstance to get
a handle on.

Test suites are notorious /bin/sh users unfortunately.

Hopefully, few packages will be as bad as gettext in this regard.
It's unclear how sustainable this approach will be going forward.

I'm also not sure how upstreamable stuff like this is either.
It's certainly not clear that this sort of "extreme prefixification"
offers enough benefits to upstream gentoo -- probably, for prefix
gentoo, it makes sense but if it can't go into mainline that could
still present a problem for gentoo-prefix inclusion.

Anyhow, we need it here so in it goes and we'll sort out the mess
later.