sys-devel/gettext: prefix -> ultra-prefix
Start walling off all of the "crazy" commits that have been going in
for "deep prefixification" under the "ultra-prefixification" moniker.

I'm doing this as a way of leaving "breadcrumbs" so that when the time
comes to bring gmt_cygwin_overlay out of "crazy town" and back to some
kind of compatibility with upstream and/or general sanity, the job won't
be intractably difficult.