sys-devel/gnuconfig: fix a bug in our bug
Sigh, this was doubly wrong:

o First, because I took everybody's word for it that I should use
  i686-pc-cygwin1.7 when in fact that was rejected upstream and
  a dead-end

o Second, because I seem to have screwed up (I wonder if it was
  my mistake or if I cut posted the rejected patch? hehe) the
  implementation of my horrible mistake with this patch as there
  was a bug in it.  Luckily, the testsuite caught it.

What is the world coming to when you can't randomly cut-paste
code into your tree from strangers on the internet without evaluating
it for correctness and expect it to magically work?

Anyhow this patch just digs us deeper into the hole I have dug
for myself by fixing the bug.  It's not worth this lengthy description,
I'm just trying to be clever I guess.  Eventually this "0003-blah" patch
(horribly named at this point) will go away but fixing it is actually
somewhat nontrivial at this point.