sys-devel/gnuconfig: ultra-prefixification
I've been patching these in other ebuilds like a drunken patchmonkey but
this time I carefully evaluated all the hard-paths and decided that almost
every single one of them is intended to capture a platform quirk of some
kind or another and therefore they shouldn't be ultraprefixified.  However
that still leaves the scripts themselves and few random spots.  Likely
this will break some other patches which is fine -- they should be removed
anyhow as I now believe I was far too gung-ho with them anyhow, plus
this is a good way to start shrinking the rediculous amount of ultraprefix
diffs I now have in-tree.  For now I'm going to consider this ultraprefixification
crap to be solved.  The patches are doing what I wanted, which is, mostly,
preventing bash-subshell-fork disasters in libtool.

My next cyglibtool hack attempt -- when I get to it -- will probably involve
attempting to prevent those subshells in the first place.  If we are lucky
this might just be enough to allow us to chuck all this ultra-prefixification
crap in the bit bucket where it very much belongs (actually it does have some
limited utility for prefix users on ancient dinosaurs with severely broken
/bin/sh, which is a real-world problem that does crop up on occasion... even
so I doubt the cost-bene adds up considering we have to keep maintaining these
gigantic piles steaming patchpoo to sustain the feature going forward).