sys-devel/libtool: VICTORY IS MINE!!!
This monster libtool update makes libtool actually work.  Unfortunately
it does so at a pretty high cost:

o ultra-prefixification of libtool
o ultra-prefixification of libtool-generated files which means that
  our overlay is not suitable for generating files intended for
  distribution
o lots of changes relative to our various upstreams, most of which
  cannot go upstream.
o slow, ugly ebuild

However, it works.  Making it "right" is now a matter of fixing
various problems derived from nasty interactions with upstream
packages.  For example, if '//' UNC notation could be deactivated,
we could get rid of all the UNC prevention hacks, or if we could
avoid rebasing problems, all of the ultra-prefixification could
go away (in fact it's worth investigating just how important
ultra-prefixification has turned out to be now that we have
solved various other problems, with an eye towards removal).

The subshell-avoidance stuff, although also ugly, is less horrible
in the sense that even outside of the "gentoo cygwin" context,
and independent of other packaging issues, it addresses a real problem
(cygwin fork() is painfully slow, not to mention prone to causing
rebase issues) which is unlikely to go away.  However much of it still
probably could not go into real libtool insofar as it relies
on bashisms or platform-specific behaviors (a big one -- I don't know
if it's really a platform-specific issue or not -- would be whether
ln -s works if no target exists for the link).  Plus some of it is
just aesthetically ugly, although seemingly "clean code" is not a
design objective that libtool values all to highly to begin with, so
maybe that's OK.

We have upgraded here to libtool-2.4.2, which we have "unmasked"
(figuratively speaking) by uncommenting the KEYWORDS clause.
There is no really good reason behind this decision -- it started as
a whimsical attempt to see how well things ported forward from 2.4-r4,
and after I accidentally blew away much of the 2.4-r4 work, I simply
decided it wasn't worth the effort to try to recreate everything.

If this causes horrible problems, so be it but my guess is this will
be a drop in the bucket compared to the horrible problems we will see
from other causes :).  Another reason I figured this was OK: I'm gambling
that by the time this overlay reaches adolescence, libtool-2.4.2  or
higher will become standard-issue anyhow.

I removed epunt_cxx since upstream did so as well, although this is not
yet reflected in portage_diffs: see bug #410877

Summary of changes:

o ultra-prefixification
o modify install-sh to avoid unneccesary subshells
o remove epunt_cxx
o use libtool-2.4.2
o libltdl: avoid unneccesary subshells
o ebuild: simulate a bootstrap when neccesary to pull changes through
  the autoconf apparati.  This is pretty ugly, but if we use the
  real bootstrap then we have maintainer-mode problems, so picked
  the lesser of two evils imo.

All tests pass!