sys-libs/ncurses: Monster 5.9 update
After lots of head scratching I eventually just gave up on 5.7 ncurses.
Figuring out what broke simply wasn't worth the effort given that my plan
was to immediately abandon it for 5.9-r1 once I got things working again.

o remove the 5.7.xxxx ebuild that we were using before and start fresh with
  the 5.9-r1 ebuild from upstream
o port forward or reimplement most of the changes from the old 5.7 build
o integrate the cygwin and "everybody else" install process as much as possible;
  this version now uses most of the upstream logic for moving libraries
  around after make install; however, I had to make significant changes to the
  upstream logic to make it apply correctly to cygwin
o Exempt cygwin from the "r" LIB_PREFIX in Ada95/configure{,in} just like
  we do for ./configure{,in}
o Somehow, the "r" LIB_PREFIX exception causes a dependency on --disable-memleak.
  That makes absolutely no sense to me, but rather than try to get to the
  bottom of the issue, I opted to just force-enable it like interix does already.
o (use-enable debug memleak) come after the interix/cyg $myconf --disable-memleak
  which means it's not being used.  One wonders if this indicates that
  it's no longer required, but anyhow, it /is/ required in cygwin ATM so I fixed
  it for both of us by moving this from the big econf statement to myconf, and
  applying the use-enable option only if we are on a system other than
  interix or cygwin.  This is definitely something that belongs upstream and is
  worthy of a gentoo bug right now (I'll try to make one soon).
o Patched progs/Makefile.in to avoid a subshell that was causing rebasing errors.
  Interestingly, this subshell cd's to ${ED}bin which is where we happen to have
  the new cygncurses.dll on cygwin -- this is highly suggestive of the possibility
  that $PWD is a frequent factor in rebasing errors in subshells, something
  I hadn't given much thought to until I noticed it.
o More?  Lots of teeny tiny changes went in; most of the changes in overlay were
  actually backports of upstream changes or reversions of changes that we formerly
  had been making here.

So far I've been too 'chicken' to emerge this into my system but it compiles and
installs like a champ, and if I discover that emerging causes some sort of disaster
I'll obviously have to deal with that (I do plan to emerge it soon).

What a challenging ebuild for cygwin!!!  The past week or two of work was basically
all moving towards making this commit work.  I guess I can't be sure all those
changes were needed but I have to say that here in ncurses I'm reasonably happy with
the results.  I still wish it weren't neccesary to make so many changes to
upstream's configure options but ncurses on cygwin is very sensitive in this regard
-- basically we are looking to hit the "sweet spot" that the cygport is using
and almost everything else is kinda broken...  At some point I should return to
this and try to reduce the number of changes but for now this is enough, finally
to start moving forward with other things.