app-shells/bash: version bump & restore cygport patch
Somehow the cygport patch had fallen out of bash-4.2_p20 -- it was
mostly just the vanilla bash that we were compiling.  This commit
gets us up to bash-4.2_p36 and ports the latest bash patches from
cygports up to that version.  The porting was quite trivial so
presumably this was just some kind of fumble.

Also adds /bin/bash.exe and siblings to CYGDLL_PROTECT in profile
since as it stands bash is not receiving cygdll protection and
this almost always triggered an insta-rebase problem.