sys-devel/gcc-config: 1.5-r2 version bump and new features
gcc-config-1.5-r2 (the current stable version) needed lots
and lots of love in order to make it do the right thing on
cygwin:

First, as with 1.4.x, there were several issues
causing //foo paths to appear which completely baffles
cygwin.

Secondly, I figured out that actually lots of the
gcc dll's were not being activated in the prefix.  Instead,
unless you happened to have /usr/i686-pc-cygwin1.7/gcc-bin
in your path (mine did although I'm not sure I remember
whether that comes from any feature in the prefix or my
custom code), you were getting those .dll's from cygwin
in "/", not from the versions portage built.  This problem
is now fixed by copying the various .dll's into /bin and
/usr/bin as appropriate (I mean, the prefixed versions of
those paths).

Thirdly, a cygwin-specific target clause
has been added to the "big case statement" in the
gcc-config-1.5 script which ensures that the correct thing
is happening with the various .dll.a libraries generated
by cygwin.

Finally, I at last got to the bottom of why I had to change
the env -i ... portageq statement in gcc-config.  The answer
was stupidly obvious: "cygwin1.dll" needs to be on the path
for anything to work.  So I was able to clean up the old code
which just kind of randomly tweaked things, and boil it down
to a more concise change relative to upstream.  Hooray!

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>