gendiffs: intelligent handling of "# $Header: ..." lines
When comparing against portage, we want gendiffs to behave
as though all the cvs header lines were empty, which is the
correct state, if I understand correctly, for the files to
have in the overlay.  This prevents lots of confusing and
uneccesary output in the generated diffs.

This also virtualizes the "---/+++" lines in a manner much
more convenient for cut/paste into actual patch files which
might be applied by somebody, and tidies up a couple other
cosmetic things as well.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>