app-arch/bzip2: sync with upstream
This eliminates the 1.0.6-r1 ebuild, which was becoming a
hassle to maintain, and updates the -r3 ebuild to contain
markers where there are indentation changes that ought
to be made.

It also rebases the cygport patch against the mingw patch
which is now applied upstream... although this is kind of
awkward since the cygport puts its own changes into mingw
which will not be used by gentoo... presumably, these
changes should at least be used for cygwin->mingw cross
compiles but for now I'm happy to ignore that issue.

Finally, it changes the cygwin makefile to hard-code BASH,
which was not working the way it used to for reasons
I can't even begin to understand.  Instead it just hard-codes
it to /bin/bash -- maybe $(EPREFIX)/bin/bash is better,
but I can't remember which comes first in bootstrap and
figured /bin/bash was probably safer now that upstream has
worked out some of the more awful rebasing issues this used
to cause.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>