sys-devel/cygwin-headers-bin: An initial shot at a cygwin headers ebuild
Yes, I know it is way sexier to build it from source, but

a) We would like to be able to bootstrap this thing somehow
b) I've got plenty more urgent stuff to do!

So for now this is a start.

Open to suggestions about how to implement slotting for these.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>