Add rebasing scripts
I use the scripts frequently to deal with rebasing problems.
Since without a full rebasing solution, cygwin-portage is not
particularly useful, I have included them here.

I have plans to improve things so that this rebasing step will
not be neccesary, but for now this is what does the trick for me.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>