scripts/{peflags,rebase}all_pfx: sync with upstream
This merges in the latest gobbledygook from cygwin upstream
to the prefix-supporting peflagsall and rebaseall scripts.

It also reverts a couple of changes I'd made second-guessing
the upstream default settings.  Hopefully this will make things
go more smoothly, 'cause they still ain't so smooth right now :(

Well, they have been worse.  Even if my second-guessing was
incredibly brilliant -- and it wasn't, trust me! -- the fact is
that people are going to run setup.exe and it's going to run
rebaseall with the default settings.  So, trying to maintain
conflicting settings will remain an uphill battle.  It turns the
rebasing issue into a wacky craps-shoot where some random thing
is going to happen depending on whether we last ran cygwin's
rebasing or prefix's -- there is more to be done but a first
step towards getting a handle on the situation is going to be
to allow these "last resort" scripts to stop fighting for
control of the rebasing strategy.

Only problem is, the auto-rebasing happening in e-builds is
still doing God-I-don't-even-remember-what in the pre-commit
phase, so now the problem is most likely that Cygwin prefix
is doing battle with itself.

Haven't decided how to deal with that but it needs to change.
Anyhow that matters is OT for this commitlog, which is clearly
a step in the right direction.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>