scripts/rebaseall_pfx: Don't count failures
When calculating the "stack" of new bases, exclude from
our calculation any files which we attempted but failed
to rebase.  The bases of these files might well have been
caclulated by gcc, some other rebasing framework, or
lord-knows-what -- anyhow they could significantly screw up
our caclulations.  In practice, this is perhaps less of an
issue than it might seem, since the files probably are
rebased when all processes are killed -- but still the
potential is there for this to be quite a problem, i.e.,
just after a cygdll-update.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>