scripts/rebaseall_pfx: Rebase perl downward like python
When rebasing perl, we now do it like python: low and growing downward;
it seems that perl is susceptible to whatever disease causes python to
give the "invalid exec format" (?) error.  I'm beginning to wonder if
libtool "dlload"ed files might be susceptible to this?

Anyhow, this seems seems better.

Also, added the scaffolding for checking the database when
determining the bases -- otherwise, by only paying attention
to those files that we rebase (and ignoring those we skip),
we are inviting collisions -- they should be mostly harmless
but we have seen that mostly harmless often isn't in practice
when it comes to these issues.

Once that is in place, it should also be possible to fully rebase
the non-prefix files, including the stuff we skipped due to basename
duplication.

The implementation itself is still unwritten.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>