eclass/multiprocessing: disable on cygwin
This temporary make-it-work patch effectively serializes
multiprocessing on cygwin.  It does so by creating noop/dummy
versions of the multiprocessing APIs on cygwin.  The reason for
doing so is that cygwin can't handle the two-way redirection
that it's using (see: http://cygwin.com/ml/cygwin/2012-07/msg00393.html)

This solution is much, much cleaner than the solution we have
in portage for exactly the same problem; if cygwin doesn't fix this
soon, we should drop the portage patch and do it this way instead.

However, note that the solution is incomplete, as I haven't yet
spent the time to grok exactly what are the semantics of
redirect_alloc_fd, so I just have it die -- so far I don't see
any consumers of that API except eclass/multiprocessing itself,
but those instances will never be reached on cygwin.  But,
IIRC, consumers do exist in portage, so, it's not just a
cut-and-paste situation there.

Also, see http://cygwin.com/ml/cygwin/2012-07/msg00384.html.
If this correct then in all probability, the next release of
cygwin will fix the problem entirely and we can just remove this
file from the overlay and un-revert the multiprocessing patches
in portage, problem solved.

So leaving well enough alone for the time being is going to be
my plan until something breaks or a new cygwin release
comes out.  If the next release fails to fix it, then I guess
I'll be forced to think about this more carefully and make
some kind of decision.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>