sys-devel/libtool: sync with upstream and sanitize
o Removes anti-maintainer mode patches, which are no longer needed
o Removes subshell-avoidance patches, which, although they would
  still have provided some benefit on cygwin, are no longer strictly
  neccesary now that the rebasing scripts work better, and required
  the horrific simulated bootstrap process in the ebuild.
o Remove ultra-prefixification, which we no longer need thanks to
  the rebasing scripts.
o Revert prefix -> prefix-gmt eprefixify migration, which got shot
  down upstream and isn't nearly as important now that we no longer
  ultra-prefixify.
o Remove the simulated bootstrap from the ebuild, which is no longer
  needed after removing all the above
o sync with upstream ebuilds

After all of the above, the ebuild patch is now quite reasonable:
we just apply a few hopefully uncontrovercial patches on cygwin,
and that's it.  Thank God, this was one of the worst ebuilds in the
overlay in terms of adding gross, non-upstreamable hacks.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>