sys-devel/libtool: Rebirth of the "subshell-avoidance" patch
Well, I gave it the 'ol college try and removed the
subshell-avoidance patches.  It didn't work out.  For example, on
my system, no amount of rebasing and screwing around could enable
me to build gettext.  That was the original problem that prompted
me to make those patches, IIRC.

So, I've respun them, this time along with the corresponding
generated files so that we can avoid reviving, along with it,
the gut-wrenching simulated bootstrap code, which always
made me want to gouge my eyes out.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>