sys-devel/gettext: parallelize autotools for cygwin
Now that we have working multiprocessing.eclass, use it for the
cygwin-specific gettext autotools processing.  Unfortunately,
we still can't just eautoreconf, so do it manually.

I'm concerned that there is a rarely triggered race-condition
here: during the libtoolize step of eautoreconf, it seems like
some of the libtool files are temporarily missing.  It seems that
other steps of the autotools process can see those missing files
and error out.

I've only seen this happen once and some things have changed since
then -- Since I can't reproduce, I'm going to go ahead and commit
this and run some smoke tests to see if I can get it to happen
again.

The fact of the matter is, having to restart this ebuild once
or even twice is still probably faster than waiting for all that
autotools crap to run.  Although I suppose if I can reproduce the
problem I'll have to consider whether it's worth keeping this patch.

Signed-off-by: Gregory M. Turner <gmturner007@ameritech.net>